+++ lib/behavior/behaviors/b_can_see_type.dart
@@ -16,7 +16,7 @@ class BCanSeeType<T extends GameComponent> extends Behavior {
   });
 
   @override
-  bool runAction(double dt, GameComponent comp, BonfireGameInterface game) {
+  bool process(double dt, GameComponent comp, BonfireGameInterface game) {
     if (comp is WithVision) {
       List<T>? list;
       comp.vision.seeComponentType<T>(
@@ -28,9 +28,9 @@ class BCanSeeType<T extends GameComponent> extends Behavior {
         },
       );
       if (list != null) {
-        return doBehavior(list!).runAction(dt, comp, game);
+        return doBehavior(list!).process(dt, comp, game);
       }
-      return doElseBehavior?.runAction(dt, comp, game) ?? true;
+      return doElseBehavior?.process(dt, comp, game) ?? true;
     } else {
       return true;
     }