+++ lib/behavior/behaviors/b_custom.dart
@@ -14,7 +14,7 @@ class BCustom extends Behavior {
     super.id,
   });
   @override
-  bool runAction(double dt, GameComponent comp, BonfireGameInterface game) {
+  bool process(double dt, GameComponent comp, BonfireGameInterface game) {
     return behavior(dt, comp, game);
   }
 }